backend/neurofin_backend/forecast/requirements.txt
flask==2.3.2
pymongo==4.4.0
pandas==2.1.2
numpy==1.26.0
tensorflow==2.15.0
python-dateutil==2.8.2
gunicorn==20.1.0
